refactor(memory): move delivery and outbox runtime modules
